source/requirements.txt
matplotlib
numpy
opencv-python
Pillow
transformers
hdbscan
mediapipe
controlnet_aux
torchvision
deepface
pytorch-lightning
insightface
onnxruntime
tensorflow